Split the variant ctor tests into separate translation units to reduce memory use
